feat enable datcite and bioschema metadata export using ids #596
